+++ dali/operators/image/convolution/laplacian_params.h
@@ -15,7 +15,7 @@
 #ifndef DALI_OPERATORS_IMAGE_CONVOLUTION_LAPLACIAN_PARAMS_H_
 #define DALI_OPERATORS_IMAGE_CONVOLUTION_LAPLACIAN_PARAMS_H_
 
-#include <numeric>
+#include <cmath>
 #include <vector>
 
 #include "dali/pipeline/data/views.h"
@@ -45,7 +45,7 @@ std::array<float, axes> GetNormalizationFactors(
   for (int i = 0; i < axes; i++) {
     auto initial = -axes - 2;
     auto exponent = std::accumulate(window_sizes[i].begin(), window_sizes[i].end(), initial);
-    factors[i] = exp2(-exponent);
+    factors[i] = std::exp2f(-exponent);
   }
   return factors;
 }
